+++ src/types/index.ts
@@ -4,8 +4,21 @@ export type BuildTool = 'maven' | 'gradle' | 'npm' | 'yarn';
 export type Runner = 'junit' | 'testng' | 'junit-cucumber' | 'testng-cucumber' | 'mocha' | 'jasmine' | 'cucumber';
 
 export interface ProjectConfig {
-    tool: AutomationTool;
-    language: ProgrammingLanguage;
-    buildTool: BuildTool;
-    runner: Runner;
+    tool: string;
+    language: string;
+    buildTool: string;
+    runner: string;
+}
+
+export interface TreeItem {
+    path: string;
+    type: string;
+    url: string;
+    children?: TreeItem[];
+}
+
+export interface PreviewProps {
+    config: ProjectConfig;
+    onClose: () => void;
+    onDownload: () => void;
 }
